# server/contracts/openapi/homeos.yaml
...
openapi: "3.0.3"
info:
  title: HomeOS API
  version: "1.0"
  description: >
    HomeOS 底座服务接口，负责身份、家庭、权限、通知、同步、搜索等基础能力。
    所有接口均需携带 JWT token（除健康检查与 JWKS 分发外）。
servers:
  - url: /api/homeos
paths:
...
  /members/snapshot:
    get:
      summary: 成员快照
      description: >
        供 authz SDK 与各服务投影使用的成员与权限快照。
        支持 If-None-Match: pver 命中 304。
        P1 期间 family_overrides 与 object_acls 恒返回空数组（㉖）。
      operationId: getMembersSnapshot
      tags:
        - internal
      security:
        - bearerAuth: []
      parameters:
        - name: fid
          in: query
          required: true
          schema:
            type: string
            format: uuid
        - name: If-None-Match
          in: header
          schema:
            type: string
      responses:
        '200':
          description: 成员与权限快照
          content:
            application/json:
              schema:
                type: object
                properties:
                  pver:
                    type: integer
                  members:
                    type: array
                    items:
                      type: object
                      properties:
                        member_id:
                          type: string
                          format: uuid
                        user_id:
                          type: string
                          format: uuid
                          nullable: true
                        name:
                          type: string
                        relation:
                          type: string
                        role:
                          type: string
                          enum: [owner, member, ward, guest]
                        avatar:
                          type: string
                          nullable: true
                  permissions:
                    type: array
                    items:
                      type: object
                      properties:
                        scope:
                          type: string
                          enum: [module, data, operation]
                        resource:
                          type: string
                        action:
                          type: string
                        condition:
                          type: string
                  family_overrides:
                    type: array
                    items: {}
                    description: P1 恒为空数组
                  object_acls:
                    type: array
                    items: {}
                    description: P1 恒为空数组
        '304':
          description: 未修改（pver 匹配）

  # === 面配置管理 ===
...
  /app/bundles:
    get:
      summary: 前端分包清单
      description: >
        按 codes 与客户端 installed 返回每面分包的 version/url/sha256/size/min_compat。
        供按需加载与远程下发使用。
      operationId: getAppBundles
      tags:
        - app
      security:
        - bearerAuth: []
      parameters:
        - name: codes
          in: query
          required: false
          schema:
            type: string
          description: 逗号分隔的 code 列表
        - name: installed
          in: query
          required: false
          schema:
            type: string
          description: 客户端已安装的分包版本 JSON
      responses:
        '200':
          description: 分包清单
          content:
            application/json:
              schema:
                type: object
                properties:
                  bundles:
                    type: array
                    items:
                      type: object
                      properties:
                        code:
                          type: string
                        version:
                          type: string
                        url:
                          type: string
                        sha256:
                          type: string
                        size:
                          type: integer
                        min_compat:
                          type: string
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT